fix!: digest consistent with the spec
Replace swid.HashEntry with a new Digest type. This need to accommodate
text algorithm IDs supported by the CoRIM spec rev. 10[1], but not allowed
by CoSWID.

Note: CoRIM borrows the definition of Digest from EAT measured
components spec[2]. The latter, unlike the CoRIM spec, defines a JSON
representation, so the JSON serialization of digests has been changed to
match that.

[2]: https://datatracker.ietf.org/doc/html/draft-ietf-rats-corim-10
[2]: https://datatracker.ietf.org/doc/html/draft-ietf-rats-eat-measured-component-12

BREAKING CHANGE: swid.HashEntry is replaced with comid.Digest; their
fields differ. Also JSON serialization for digests changed to be
consistent with the one defined for eatmc.digest.

+++ comid/measurement.go
@@ -651,7 +651,7 @@ func (o *Measurement) SetMinSVN(svn uint64) *Measurement {
 // AddDigest add the supplied digest - comprising the digest itself together
 // with the hash algorithm used to obtain it - to the measurement-values-map of
 // the target measurement
-func (o *Measurement) AddDigest(algID uint64, digest []byte) *Measurement {
+func (o *Measurement) AddDigest(algID int, digest []byte) *Measurement {
 	if o != nil {
 		ds := o.Val.Digests
 		if ds == nil {
